drivers: wifi: nrf71: Move util sources into common module
Relocate util.c and util.h from utils/ to src/common and inc/common,
update includes to common/util.h, and drop the utils directory.
