+++ drivers/wifi/nrf71/osal/fw_if/umac_if/src/common/fmac_cmd_common.c
@@ -14,7 +14,7 @@
 
 #include <nrf71_wifi_ctrl.h>
 #include "common/fmac_structs_common.h"
-#include "common/fmac_util.h"
+#include <common/util.h>
 #include <zephyr/logging/log.h>
 
 LOG_MODULE_DECLARE(wifi_nrf, CONFIG_WIFI_NRF71_LOG_LEVEL);
