+++ drivers/wifi/nrf71/osal/fw_if/umac_if/src/radio_test/fmac_event.c
@@ -14,7 +14,7 @@
 
 #include <nrf71_wifi_ctrl.h>
 #include "radio_test/fmac_structs.h"
-#include "common/fmac_util.h"
+#include <common/util.h>
 #include <zephyr/logging/log.h>
 
 LOG_MODULE_DECLARE(wifi_nrf, CONFIG_WIFI_NRF71_LOG_LEVEL);
