+++ drivers/wifi/nrf71/osal/fw_if/umac_if/src/system/fmac_event.c
@@ -21,7 +21,7 @@
 #include "system/fmac_peer.h"
 #include "system/fmac_ap.h"
 #include "system/fmac_event.h"
-#include "common/fmac_util.h"
+#include <common/util.h>
 #include <zephyr/logging/log.h>
 
 LOG_MODULE_DECLARE(wifi_nrf, CONFIG_WIFI_NRF71_LOG_LEVEL);
