+++ drivers/wifi/nrf71/osal/fw_if/umac_if/src/system/fmac_vif.c
@@ -13,7 +13,7 @@
 #include "system/fmac_vif.h"
 
 #include <nrf71_wifi_ctrl.h>
-#include "common/fmac_util.h"
+#include <common/util.h>
 #include <zephyr/logging/log.h>
 
 LOG_MODULE_DECLARE(wifi_nrf, CONFIG_WIFI_NRF71_LOG_LEVEL);
