+++ drivers/wifi/nrf71/src/common/util.c
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2025 Nordic Semiconductor ASA
+ * Copyright (c) 2026 Nordic Semiconductor ASA
  *
  * SPDX-License-Identifier: BSD-3-Clause
  */
@@ -12,7 +12,13 @@
 #include <common/mem_mgmt.h>
 #include <string.h>
 #include <zephyr/sys/util.h>
-#include <util.h>
+#include <common/util.h>
+#include <common/fmac_structs_common.h>
+#include <nrf71_wifi_ctrl.h>
+#ifdef NRF71_SYSTEM_MODE
+#include <common/nbuf_mgmt.h>
+#include <system/fmac_api.h>
+#endif /* NRF71_SYSTEM_MODE */
 #include <zephyr/logging/log.h>
 
 LOG_MODULE_DECLARE(wifi_nrf, CONFIG_WIFI_NRF71_LOG_LEVEL);
@@ -155,4 +161,92 @@ unsigned char nrf_wifi_utils_get_op_band(void)
 		return NRF_WIFI_OP_BAND_2GHZ | NRF_WIFI_OP_BAND_5GHZ;
 	}
 	return NRF_WIFI_OP_BAND_2GHZ | NRF_WIFI_OP_BAND_5GHZ | NRF_WIFI_OP_BAND_6GHZ;
+}
+
+bool nrf_wifi_util_is_multicast_addr(const unsigned char *addr)
+{
+	return (0x01 & *addr);
+}
+
+bool nrf_wifi_util_is_unicast_addr(const unsigned char *addr)
+{
+	return !nrf_wifi_util_is_multicast_addr(addr);
+}
+
+bool nrf_wifi_util_ether_addr_equal(const unsigned char *addr_1,
+				    const unsigned char *addr_2)
+{
+	const unsigned short *a = (const unsigned short *)addr_1;
+	const unsigned short *b = (const unsigned short *)addr_2;
+
+	return ((a[0] ^ b[0]) | (a[1] ^ b[1]) | (a[2] ^ b[2])) == 0;
+}
+
+unsigned short nrf_wifi_util_rx_get_eth_type(void *nwb)
+{
+	unsigned char *payload = (unsigned char *)nwb;
+
+	return payload[6] << 8 | payload[7];
+}
+
+unsigned short nrf_wifi_util_tx_get_eth_type(void *nwb)
+{
+	unsigned char *payload = (unsigned char *)nwb;
+
+	return payload[12] << 8 | payload[13];
+}
+
+enum nrf_wifi_status nrf_wifi_check_mode_validity(unsigned char mode)
+{
+	if ((mode ^ NRF_WIFI_STA_MODE) == 0) {
+		return NRF_WIFI_STATUS_SUCCESS;
+	}
+#ifdef NRF71_RAW_DATA_RX
+	else if ((mode ^ NRF_WIFI_MONITOR_MODE) == 0) {
+		return NRF_WIFI_STATUS_SUCCESS;
+	}
+#endif /* NRF71_RAW_DATA_RX */
+	return NRF_WIFI_STATUS_FAIL;
+}
+
+bool nrf_wifi_util_is_arr_zero(unsigned char *arr,
+			       unsigned int arr_sz)
+{
+	for (unsigned int i = 0; i < arr_sz; i++) {
+		if (arr[i] != 0) {
+			return false;
+		}
+	}
+
+	return true;
+}
+
+#ifdef NRF71_SYSTEM_MODE
+unsigned char *nrf_wifi_util_get_ra(struct nrf_wifi_fmac_vif_ctx *vif,
+				    void *nwb)
+{
+	if ((vif->if_type == NRF_WIFI_IFTYPE_STATION)
+#ifdef NRF71_RAW_DATA_TX
+	    || (vif->if_type == NRF_WIFI_STA_TX_INJECTOR)
+#endif /* NRF71_RAW_DATA_TX */
+#ifdef NRF71_PROMISC_DATA_RX
+	    || (vif->if_type == NRF_WIFI_STA_PROMISC)
+	    || (vif->if_type == NRF_WIFI_STA_PROMISC_TX_INJECTOR)
+#endif
+	    ) {
+		return vif->bssid;
+	}
+
+	return nrf_wifi_nbuf_data_get(nwb);
+}
+#endif /* NRF71_SYSTEM_MODE */
+
+void *wifi_fmac_priv(struct nrf_wifi_fmac_priv *def)
+{
+	return &def->priv;
+}
+
+void *wifi_dev_priv(struct nrf_wifi_fmac_dev_ctx *def)
+{
+	return &def->priv;
 }
