+++ drivers/wifi/nrf71/src/common/util_shell.c
@@ -12,7 +12,7 @@
 #include <zephyr/kernel.h>
 #include <zephyr/sys/sys_heap.h>
 #include <common/fw_if/nrf71_wifi_ctrl.h>
-#include <common/fmac_util.h>
+#include <common/util.h>
 #include <system/fmac_api.h>
 #include <common/mem_mgmt.h>
 #include <system/core.h>
