+++ drivers/wifi/nrf71/src/radio_test/main.c
@@ -15,8 +15,8 @@
 #include <zephyr/logging/log.h>
 #include <zephyr/device.h>
 
-#include <util.h>
 #include <common/rf_params.h>
+#include <common/util.h>
 #include <radio_test/core.h>
 
 #define DT_DRV_COMPAT nordic_wlan
