+++ drivers/wifi/nrf71/src/system/main.c
@@ -28,10 +28,9 @@
 #endif /* CONFIG_NET_CONNECTION_MANAGER_CONNECTIVITY_WIFI_MGMT */
 
 #include <common/status.h>
-#include <util.h>
-#include <common/fmac_util.h>
 #include <common/rf_params.h>
 #include <common/vtf.h>
+#include <common/util.h>
 #include <system/core.h>
 #include <drivers/wifi/nrf71/nrf71_wifi_coex.h>
