+++ drivers/wifi/nrf71/src/system/wpa_supp_if.c
@@ -20,7 +20,7 @@ LOG_MODULE_DECLARE(wifi_nrf, CONFIG_WIFI_NRF71_LOG_LEVEL);
 #include <net_private.h>
 
 #include <system/core.h>
-#include <common/fmac_util.h>
+#include <common/util.h>
 #include <system/wifi_mgmt.h>
 #include <system/wpa_supp_if.h>
 #include <system/fmac_peer.h>
